Move ClaimsListDao related classes to new package
